service-whatsapp-parser/requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
python-multipart
pillow